refactor: split up `src/locale` folder
Move en/ru locale JSON files to root locale folder
Move i18n config to root
Move `useDictionary` hook to `src/hooks`
Move `getDictionary` server helper to `src/helpers`
Move `DictionaryTypes` to `src/types`
Move `dictionaries` constant to `src/constants`
