+++ src/app/[locale]/_components/CategoriesGrid/interfaces.ts
@@ -1,4 +1,4 @@
-import { type DictionaryType } from "@/locale/get-dictionary";
+import { type DictionaryType } from "@/types/dictionary";
 
 export interface CategoriesGridProps {
   locale: PageLocaleParams["params"]["locale"];
