+++ src/app/[locale]/_components/FeaturedInBlock/interfaces.ts
@@ -1,4 +1,4 @@
-import { type DictionaryType } from "@/locale/get-dictionary";
+import { type DictionaryType } from "@/types/dictionary";
 
 export interface FeaturedInBlockProps {
   dictionary: DictionaryType["featuredIn"];
