+++ src/app/[locale]/_components/PostsBlock/FeaturedPostBlock/interfaces.ts
@@ -1,4 +1,4 @@
-import { type Locale } from "@/locale/i18n-config";
+import { type Locale } from "i18n-config";
 
 export interface FeaturedPostBlockProps {
   locale: Locale;
