+++ src/app/[locale]/_components/PostsListCard/interfaces.ts
@@ -1,5 +1,6 @@
+import { type Locale } from "i18n-config";
+
 import { type Post } from "@/constants/data/posts";
-import { type Locale } from "@/locale/i18n-config";
 
 export interface PostsListCardProps {
   locale: Locale;
