+++ src/app/[locale]/_components/TestimonialsBlock/interfaces.ts
@@ -1,4 +1,4 @@
-import { type DictionaryType } from "@/locale/get-dictionary";
+import { type DictionaryType } from "@/types/dictionary";
 
 export interface TestimonialsBlockProps {
   dictionary: DictionaryType["testimonials"];
