+++ src/app/[locale]/_components/layout/LocaleSwitcher/index.tsx
@@ -1,12 +1,12 @@
 "use client";
 
 import clsx from "clsx";
+import { i18n } from "i18n-config";
 import Link from "next/link";
 import { memo } from "react";
 
 import { useLocaleSwitcher } from "@/app/[locale]/_hooks";
 import { Body } from "@/components/ui";
-import { i18n } from "@/locale/i18n-config";
 
 import { type LocaleSwitcherProps } from "./interfaces";
 import styles from "./LocaleSwitcher.module.scss";
