+++ src/app/[locale]/_components/layout/SubscribeBlock/interfaces.ts
@@ -1,3 +1,3 @@
-import { type DictionaryType } from "@/locale/get-dictionary";
+import { type DictionaryType } from "@/types/dictionary";
 
 export type SubscribeBlockProps = DictionaryType["subscribe"];
