+++ src/app/[locale]/_components/skeletons/LocaleSwitcherSkeleton/index.tsx
@@ -1,5 +1,6 @@
+import { i18n } from "i18n-config";
+
 import { Body } from "@/components/ui";
-import { i18n } from "@/locale/i18n-config";
 
 export default function LocaleSwitcherSkeleton() {
   return (
