+++ src/app/[locale]/about-us/_components/AboutUsHeader/MissionVision/interfaces.ts
@@ -1,4 +1,4 @@
-import { type DictionaryType } from "@/locale/get-dictionary";
+import { type DictionaryType } from "@/types/dictionary";
 
 export interface MissionVisionProps {
   dictionary: DictionaryType["aboutUsPage"]["missionVision"];
