+++ src/app/[locale]/about-us/_components/AboutUsHeader/OverviewBlock/interfaces.ts
@@ -1,4 +1,4 @@
-import { type DictionaryType } from "@/locale/get-dictionary";
+import { type DictionaryType } from "@/types/dictionary";
 
 export interface OverviewBlockProps {
   dictionary: DictionaryType["aboutUsPage"]["overview"];
