+++ src/app/[locale]/about-us/_components/AboutUsHeader/interfaces.ts
@@ -1,4 +1,4 @@
-import { type DictionaryType } from "@/locale/get-dictionary";
+import { type DictionaryType } from "@/types/dictionary";
 
 export interface AboutUsHeaderProps {
   headerDictionary: DictionaryType["aboutUsPage"]["header"];
