+++ src/app/[locale]/author/[id]/_components/Posts/interfaces.ts
@@ -1,4 +1,4 @@
-import { type Locale } from "@/locale/i18n-config";
+import { type Locale } from "i18n-config";
 
 export interface PostsProps {
   locale: Locale;
