+++ src/app/[locale]/blog/[id]/_components/SimilarPosts/PostCard/interfaces.ts
@@ -1,5 +1,6 @@
+import { type Locale } from "i18n-config";
+
 import { type Post } from "@/constants/data/posts";
-import { type Locale } from "@/locale/i18n-config";
 
 export interface PostCardProps {
   locale: Locale;
