+++ src/app/[locale]/blog/_components/FeaturedPostHeader/interfaces.ts
@@ -1,4 +1,4 @@
-import { type Locale } from "@/locale/i18n-config";
+import { type Locale } from "i18n-config";
 
 export interface FeaturedPostHeaderProps {
   locale: Locale;
