+++ src/app/[locale]/category/[id]/_components/TagsSearch/interfaces.ts
@@ -1,4 +1,4 @@
-import { type DictionaryType } from "@/locale/get-dictionary";
+import { type DictionaryType } from "@/types/dictionary";
 
 export interface TagsSearchProps {
   dictionary: DictionaryType["categoriesPage"]["search"];
