+++ src/app/[locale]/contact-us/_components/ContactForm/interfaces.ts
@@ -1,4 +1,4 @@
-import { type DictionaryType } from "@/locale/get-dictionary";
+import { type DictionaryType } from "@/types/dictionary";
 
 export type ContactFormProps = {
   formLocale: DictionaryType["contactUs"]["form"];
