+++ src/components/ui/AuthorLink/interfaces.ts
@@ -1,5 +1,6 @@
+import { type Locale } from "i18n-config";
+
 import { type Author } from "@/constants/data/authors";
-import { type Locale } from "@/locale/i18n-config";
 
 export interface AuthorLinkProps extends Pick<Author, "id" | "name"> {
   locale: Locale;
