+++ src/helpers/getFormattedDate.ts
@@ -1,5 +1,6 @@
+import { type Locale } from "i18n-config";
+
 import { monthMap } from "@/constants/date";
-import { type Locale } from "@/locale/i18n-config";
 
 export function getFormattedDate(date: Date, locale: Locale) {
   const monthText = monthMap[locale][date.getMonth()]!;
