+++ src/middleware.ts
@@ -1,9 +1,8 @@
 import { match as matchLocale } from "@formatjs/intl-localematcher";
+import { i18n } from "i18n-config";
 import Negotiator from "negotiator";
 import type { NextRequest } from "next/server";
 import { NextResponse } from "next/server";
-
-import { i18n } from "./locale/i18n-config";
 
 function getLocale(request: NextRequest): string | undefined {
   const negotiatorHeaders: Record<string, string> = {};
