+++ src/types/dictionary.ts
@@ -1,0 +1,3 @@
+import { type getDictionary } from "@/helpers/getDictionary";
+
+export type DictionaryType = Awaited<ReturnType<typeof getDictionary>>;
