+++ types/app.d.ts
@@ -1,4 +1,4 @@
-import { type Locale } from "../src/locale/i18n-config";
+import { type Locale } from "i18n-config";
 
 declare global {
   declare type PageLocaleParams = { params: { locale: Locale } };
